+++ runSystemTests.sh
@@ -22,5 +22,5 @@ yes | /Users/$USER/Library/Android/sdk/cmdline-tools/latest/bin/sdkmanager --lic
 ## command to run specific test (with regex) in a loop (can be used to verify flaky test is stable)
 #for i in {1..50}; do
 #  ./gradlew :collector-bitmovin-player:cleanPixel6api30DebugAndroidTest
-#  ./gradlew :collector-bitmovin-player:pixel6api30DebugAndroidTest -Pandroid.testInstrumentationRunnerArguments.tests_regex=com.bitmovin.analytics.bitmovin.player.PhoneBasicScenariosTest.test_live_playWithAutoplayAndMuted || exit
+#  ./gradlew :collector-bitmovin-player:pixel6api30DebugAndroidTest -Pandroid.testInstrumentationRunnerArguments.tests_regex=com.bitmovin.analytics.bitmovin.player.PhoneBasicScenariosTest.test_wrongAnalyticsLicense_ShouldNotInterfereWithPlayer || exit
 #done