+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Development
 
 ### Changed
-- Upgrade example app to latest bitmovin player (v3.37.2)
+- Upgrade example app to latest bitmovin player (v3.38.0)
 
 ### Added
 - [internal] `delayed` property to `EventData` and `AdEventData` to indicate potentially longer delay between the actual event time and send time