+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Development
 
 ### Changed
-- Upgrade example app to latest bitmovin player (v3.38.0)
+- Upgrade example app to latest bitmovin player (v3.38.0) and latest exoplayer version (v2.18.7)
 
 ### Added
 - `CollectorConfig.longTermRetryEnabled` with default value of `false`. When set to `true`, analytics events that have failed to be sent are cached in a persistent way and re-sent as soon as there is network again